requirements.txt
langchain>=1.2.15
langchain-openai>=1.1.16
langchain-community>=0.4.1
langgraph>=1.1.9
tavily-python>=0.3.3
python-dotenv>=1.0.1
